Add files via upload
Insert a Node at a Specific Position in a Linked List:
This method inserts the node at the given index in the linked list. n this method, we create a new node with given data , a current node that equals to the head, and a counter ‘index’ initializes with 0. Now, if the position is equal to zero it means the node is to be inserted at begin so we called insert At Begin() method .else we run a while loop until the current node is not equal to None . we increment the position by 1 and make the current node next of it. If current node is equal to None it means that the index is not present in the list and we print “Index not present”.